move http types to versioned dir
