+++ internal/transformer/transaction.go
@@ -1,7 +1,7 @@
 package transformer
 
 import (
-	httptransaction "github.com/hampgoodwin/GoLuca/pkg/http/transaction"
+	httptransaction "github.com/hampgoodwin/GoLuca/pkg/http/v0/transaction"
 	"github.com/hampgoodwin/GoLuca/pkg/transaction"
 	"github.com/hampgoodwin/errors"
 )
